Added a "faction" var to mobs. This var gets checked when hostile simple animals attack something. Default is "neutral."
Changed the hostile code to allow for ranged mobs. Will port syndies over to hostile at some point (I don't want to go and replace the syndies on all the away maps right now).

Added pirate simple animals for Inti (icons are from him). Icons for russians and an on mob bear belt are in as well. There is a ranged version and a melee version.

Set factions for constructs, shades, carp, bears, faithless.


git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5176 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 338b934160f9e89c6fe4fae2279c935a982252f8