Runtime fixes:
Fixed a runtime caused by someone holding an input window open with the RD Core Computer.

Fixed a runtime caused by an area being null... I have no idea how this is even possible. Tracking the source of the problem will take ages.

Added a check for range and source in hear(). Again.. I'm not sure how this ever even fails. This may also fix a runtime in say() code where it's trying to compare a list of objects to the character "0".

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5282 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: cab176618a03acc6a2a697a160d835c2e9280cb8