Made golems a bit faster so they're not so godawful to play.
Made the adamantine core create a golem creation rune (and sends a notice to all ghosts that a rune has been created in whatever location)

The rune lights up when a ghost is hovering over it.

When someone clicks the rune, it'll pull a ghost hovering over it to make the golem instead of a random one.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5558 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 3d076264bd9fe19ec05f05ee3be57d5bb5e27e12