Fixes an issue in grammar!
Specifically, the way that it was coded made it so that the word
the would appear twice. There... really wasn't a way to fix that
in a way that would effect a lot more things, that I could find.
As a result, I simply reworded it so that it can work with or without
the word, 'the'.

closes #2771


Former-commit-id: adde619a06b705fe8b9f4b0c688cfb83a2084e33