Adds Uranium Capability to Circuit Imprinters
Well, looking over the RUST circuits... They required uranium
which we didn't even have implemented in the code to work with
circuit imprinters. I considered simply changing them to use
a different resource, but it honestly makes more sense to improve
the current code to have it more versatile-- To this end, I've made
circuit imprinters capable of using uranium.

As a further note...

closes #2757


Former-commit-id: 0b20496f7df852f0c34e7697e7a32ea8e35dbc0a