Separated ERT from red alert and other small changes...
Separated the ERT to be callable via the keycard auth device.

Gave the ERT call different messages to work with the above, as well as a message when it has been denied.

The old base amount of 50 was way too stronk. Furthermore, it was added to by other factors from the game. It has been lowered a bit and a proc added in to increase it the longer the round has been going on.

Made it so that you may only attempt to call the ERT a single time.

Changed it so that you only have 5 minutes to join the ERT after it has been called.


Former-commit-id: 3e364635b273e3c22eb83c95067e0f8a87dd0b73