Changes to the radiation storm event...
It now opens up maintenance access to all people upon start. Furthermore,
one minute after ending (and with a warning), it will return maintenance
requirements to how they were before the event. In addition to this, it
has been buffed to closer to the original amount of damage for the event.


Former-commit-id: a21b7deb1f361a9bd00cbbfff97b036120454a6e