Medical Hoopery
Added eyes as an internal organ, removed the eye-stat var on mobs and changed it to properly direct it to eye damage. Altered all reference.

Updated the sleeper to allow for ejection of the beakers and occupants in panel.

Updated the body scanner to check for prosthetics/mechanized organs, and inform the individual of the blind disability.

Laid the groundwork for mechanical and assisted internal organs.

Groundwork for chemical: Peridaxon, a chemical that allows for the slow regeneration from organ damage.


Former-commit-id: cba1a9585db841cbf89506e0b87da08ed293dcc4