Bugfix(2126) Traitors still have loyalty implants, admins should be able to remove/disable implants
Added a proc is_loyalty_implanted() that returns 1 if so, and 0 if not.

Added the ability to remove and add loyalty implants from the traitor panel.

Changed the checks on antags to not care if they are section heads, but to instead check for a loyalty implant.

Fixed a minor bug where traitorborgs made through traitor panel was not adding law 0

Adding a loyalty implant to an antagonist through traitor panel will also remove them from being an antag.


Former-commit-id: 9ddbc905e9c3d69fee114f4085938418f8ec975f