Tanks will also keep up with who manipulated them last. Tanks placed in canisters, scanned with PDA or other analysis tool will set the manipulated_by variable to whoever did so.
The idea is that this person is aware what's in the tank and will act accordingly when you press the internals HUD icon.

If the user was the last person to manipulate the tank (or it's not been manipulated) then the user is smart about selecting the proper tank.

If the user was not the last person to manipulate the tank then he is only checking the pressure gauge to see what tank is fullest that matches his species breath_type

If that tank has nasties in it, welp he is going to get a breathful of nasties.


Former-commit-id: 52d328623a46dbcb844a502a29881f9b8e4d666e