Teleport bug fix
Since we're unable to reproduce this bug in a testing environment and I've already concluded this isn't a telescience bug then I am leaning more towards Xenoarchiology, specifically the artifacts.

This sanity check basically makes it where if there is an artifact on null terf or inside a container it won't process the effect.
If it's the teleport effect of artifacts that is causing this to happen and it's based on the item being in nullspace somehow then it should squash the bug.
Though the root of this is probably #4037 I cannot reproduce the runtime either.

So here goes, a stab in the dark.


Former-commit-id: 4cb9cd0911e9dda3bb7a735140e5823a772791f7