Vending Machine PIN fix.
Removed requirement for PIN on vending machines with an easy way of adding it back in the future
or setting machines to not use it or use it depending on what we decide later on.

Someone attempted to make it where it would accept any PIN and this is just as annoying as typing in a PIN.

So...space communism, swipe your card and it spits out an item.  No PIN prompt at all.


Former-commit-id: 8e518cb4a9fe29fc23811d91bbf9be3e5c920a2f