PDA NanoUI template fixes.
Cartrdige name is no longer fixed size causing oddities with really long names (quartermasters).

Fixed a </span> to not have a line feed in the middle of it making the UI unusable in mode 47 if there are no requests.

Set the Supply Records to use the same style as Security/Medical Records.


Former-commit-id: 630fe9af01799c7665dbd6a7d307e2a7699d6357