Refined earlier fix
Tried a couple of other solutions, but this ended up being the most reliable.

Now checks all areas associated with the firedoor to see if they should still be down.

Also makes a final check before closing the door in case the fire alarm turns off in the middle of an autoclose cycle.


Former-commit-id: e145b616810fd1a04fd23fa69146e76bd9ef8d84