Code effeciency project: handle_embedded_objects
Before:  EVERYTIME someone moved their character this proc would be called and loop through every organ looking for implanted items to see if it needs to apply damage.

After:  We create a flag that is set when an item embeddes, only when that flag is set do we do that loop through organs upon movement.  Also every 10 ticks while that flag
is set we will check to see if we need to unset it.

Former-commit-id: b392ab8e0e11a8130385e48292572d3294d8a03b