Code effeciency project:  Canister icon updates
Before:  If a canister is hooked up to a connector port it's icons were being deleted and recreated every tick.

After:  We store an update_flag that is a bitfield of the current icon and overlay states, and created a proc to check if it should change.  If not, we don't mess with icons.  Real simple.

Former-commit-id: 56f61d2296adac1e5f20be2c9bf10931bee6e634