Compile fix / optimizations
Removed client check from mob so that if a pAI is being held by a SSD client it can still hear.

Removed 3rd deep .loc check 2 is plenty, and was making the proc rather expensive
(if you shove a pAI in a bag and toss the bag into a locker, the pAI will be deaf, you jerk).


Former-commit-id: 04a0537b1fbc9f4d2a923751a0331ff9e527f1b9