Engineering areas fix
Previously I had the engineering and atmos monitoring rooms sharing the same area. To separate the air vents I changed the air vent and air alarm frequencies. This works, but could be a little confusing to some when there are air vents listed by the air alarms that don't work (since air alarms don't filter out vents on different frequencies). This splits the rooms into separate areas to avoid unnecessary confusion.

This was also the case with the hallway outside the engine monitoring room.


Former-commit-id: 6622bbb89668e18ef18c7792571e5c7f799e402c