Code effeciency project:  Ore processing.
Before:  The stacker and processor used del(item) to get rid of the items causing /obj/movable/del() to spike up and cause fun amounts of lag.  Also we have seperate variables for if a certain ore type is selected in the processing machine and the cooking portion of process() uses a huge string of if()'s which check 8 or so variables and is terrible and probably a pain in the ass should someone decide to later add ore types and recipes.

After:  We just set item.loc = null when it goes in the stacker of processor and garbage collection takes care of it within a tick.  Those different variables for ore selection are now just one variable which is a bitfield, much easier to read, much easier to expand upon.

Former-commit-id: 7ce4073135fe3bcb6809974741afbf34206ec11b