VV drop down menu to resend NanoUI resource files.
Some players are having issues with NanoUI not opening up or acting funny
before the only fix was to have them log out, and log back in, check if that worked.
If not log out, delete cache log back in, if that don't work not much can be done.

Now we just skip the whole logout login thing and force the server to send the files
to their client once again skipping the need for their character to take a dirtnap for a moment.


Former-commit-id: 0e369578cc7ebdfe94049def45da5ccb450606a7