Code effeciency project:  Handle_Organs (human/life())
We do a bit of pre-processing to see if we want to do checks on all the organs.

First we don't even do preprocessing unless we have damage of some kind
afterwards we process what is hurt until it gets better or overall we get worse
when we go ahead and check everything else out again.

Can add a safety check in life every 30 ticks or so go ahead and check every organ again but I don't think it's needed.


Former-commit-id: f1506acf1ec3efcd2741ae71c3a0f8e7b7eb08e5