Deadlier supermatter.  If you don't add coolant, you will have a bad time.
Before:  Supermatter would raise the temperature without taking into account how many moles it grabbed in it's "breath" meaning it was better that you had a lower pressure inside your engine chamber so the engine would spit back out it's "breath" and it couldn't heat up the engine room as much.

Now:	Supermatter will take into account the heat_capacity of the breath it takes and will heat up it up accordingly and proportional to how much energy it has.

Gone are the days you could run the supermatter engine without adding coolant.

Also:  Added a debugverb that setups the engine for you, the verb is based off of baystations map and areas and shouldn't require too much change to bring it in line for other servers.

Former-commit-id: 44578b78324670653d8bb77d891d6b0a939eefbe