Accessories are now visible when attached
Uses overlays.

I was originally planning to use the SOUTH direction of the
icons/mob/ties.dmi image as the overlay, however this had problems with
the fact that those images are directional. Rather than override Move()
in clothing/under to reset dir to SOUTH whenever clothing with an
attached accessory was thrown or dragged, I thought it would be less
hacky to read the overlays from a separate .dmi.


Former-commit-id: 6e391c01c979814241a5031629ed548d7071d336