Docking overrides now have a clear purpose
Originally the purpose of the docking override was to allow the shuttle
to move even if it couldn't undock properly (e.g. the doors were stuck
open), however it makes more sense just to have the shuttle take care of
that. This commit now makes the behaviour of the docking overrides line
up with it's new purpose.


Former-commit-id: acec01fb1415d9b6d6cfc625baec06b41b17644f