Fixes #5261
Replaces the super strength full-tile reinforced windows on the research outpost with standard health variants.
Did consider renaming the super-variants to avoid accidental use in the future but I thought not after seeing the massive diff it'd cause.

Should be fully compatible with PR #5363 no matter the merge order but should be safest to merge #5363 first, then this.


Former-commit-id: d3e1b3a97f2cb67298ea5fb9277a66135cdb1adb