Partially fixes #5422.
The button to activate the flash in the permabrig is now properly assigned with an id.
Would previously activate all un-tagged mounted/stationary flashes in the known universe.

The door button is still not functional, appears to be mapped as correctly as can be but I don't think doors (with access requirements) can currently be remote controlled the same way as shutters.


Former-commit-id: af34a8d4103ba63987f4b32415307d387fea7ffe