Makes it possible for the AI to alt+left click turfs in camera view and grants ctrl/alt shortcuts to the turret control.
AI can now alt+left click turfs in camera view to list and interact with objects in the status tab. For example allows for interacting with emergency shutters beneath grills.
alt-clicking doors now notifies the AI when they are electrified/un-electrified as, unlike most other shortcuts, this doesn't have a visual cue.
When interacting with doors and the AI-control is disabled hacking is now automatically initiated, as if the AI had attempted to open the door 'menu'.
When borgs/AIs try to interact with an emagged door they now receive a feedback message that it's unresponsive.

Parts of implementation done by porting code from tg-station.


Former-commit-id: d02e483110650012c7436fa7a0f80d44c0f63671