General map changes
Sneaky second CE-door returned. It's gone now.
Moved disposal-to-space sign in Xenoflora to the disposal unit.
The doors of the emergency storages on the starboard side of the station no longer imply that they are maintenance sections but, indeed, emergency storages.
Added oil, webs, dirt, and clutter randomly about the maintenance tunnels.


Former-commit-id: 4ce3f11aa3f1435e6e8d77eb633d91ada78fb316