Fix for rad suits not protecting against radiation
Rad suits were not providing complete protection from radiation. This is
probably not the best solution. Better solutions would be to either add
the HANDS and FEET coverage flags as done here in addition to updating
the sprite to reflect the additional coverage, or to add radiation proof
gloves and boots as part of the radiation suit set (potentially
problematic for taj and unathi engineers).


Former-commit-id: c8f2d040a1936e594a03f160aeb182d53c3fba59