Add round start check to Declare Ready
Added a check to 'declare ready' to prevent readying up after the round
started. Allowing this to happen caused a bug where the player would be
unable to change their character setup if they declared ready after the
round started, such as to look at the manifest to see which character
would be best to play, until after reconnecting.


Former-commit-id: 4c5665d4fd51b03fbc6d62c32a8b2210ffd306a8