PDA newscaster
One can one access and view newsfeeds from the comfort of one's PDA.
Possible to change and disable the newsmessage beep.

Known issues:
Unlike browse() NanoUI does not appear to wait for browse_rsc(), the method which to transfers image to the client's cache, to finish.
This often results in newsfeeds not always displaying their images the first time they are viewed.


Former-commit-id: 9c41148a08e568e04e9341c3174242462ee63f5e