Fixes #6951
Replaces two standard SMES instances on station, five instances on the asteroid.
Adds an engine monitor screen to the CE's office. Allows the CE to check the engine from the comfort of his office and to allow one to be sure that the ejection vent is open before ejecting.
Makes the engineering showers prettier by having shower curtains display above windows.


Former-commit-id: 86258b6e8a7a92a751c67c380308cb233a2471f3