Newscaster and photo cleanup
Synthetics can no longer magically print color images on a black-and-white photocopier.
No longer possible for organics to acquire a synth-image by un-attaching it from a newscaster.
News articles now come with a timestamp and photo captions, if anything was penned on an attached photo.
Reception-checks no longer unconditionally sleeps the thread. Was causing havoc with NanoUI.

Misc. code refactoring.


Former-commit-id: 77c52f48a8aecf986f2aa241785ffde403ffae36