Adds 'me' hotkey, updates hotkeys-help
It bothered me that there was no hotkey for emote that wasn't on the
function row. Y is now me. Activate-held-item is now only Z instead of Z
and Y.

The hotkeys-help verb in OOC tab is changed to include the changes
above, and the previously unmentioned function row hotkeys that are
pertinent to players. (Admin verbs are unmentioned)


Former-commit-id: 8c83d4d51df3b9a93f27d3ef8fc441169df7b836