Lowered SMES coil price to 75
It is a valuable piece of technology, but alone it is not really powerful at all. This is the same price as the RUST core and TE generator, which are both admittedly much more complicated. I know this was for balance reasons, but as it is now engineers will rather steal the coils from solar panel SMESes than to deal with cargo who are not very supportive of ordering expensive stuff like this.

Former-commit-id: 2ca4ccd23abec481d461778593f22dc64c445424