Surgical kits can now contain their equipment.
The surgical kit can now contain the equipment it came with. It can also not contain more items of a given type than it was spawned with.
This means that even if the storage container otherwise has sufficient space it can, for example, still only ever contain 1 surgical saw.

Fixes #11298. Fixes #11297.


Former-commit-id: fd03e4d59b63d90c11ca586f89b34461fece75fa