Rotational icons.
/icons/obj/waterclosets.dmi - sink(_alt)
sink - North facing sink adjusted to be aligned with the wall, when the other directions are
sink_alt - Adjusted to the best of my abilities to have the west/east facing directions be aligned against with the wall when the south facing direction is. However, some mapping pixel shifting will still be needed.


Former-commit-id: 4254289dec2673261207d661fcb3e205f76fa252