Preferences now store client key instead of the client itself.
For reasons unbeknown to me, the client var is reset when the client logs outs and thus the preference mob cease to be.
Now stores the client key, and acquires the appropriate mob based on that, if possible.
Also removes the sanitation check because the holder isn't set before the preferences are created.

Fixes #12341.


Former-commit-id: 19b7ebf047f3c9492fe34e3a523f96ef90429ef8