Properly log dropper-related assault for admins
Fixes bug #1219

Moving the logging from post-dropping (when the dropper is empty) to Pre-drop, post-dropable check (it should not be located at the point when a drop is going to occur, but just prior to actually doing so).

As always, please review this change to make sure I didn't break anything.  All I did was cut-paste the lines to a few lines above, but hey, never hurts to double-check!

Former-commit-id: 3a254b5aa765b712b6586d7c067ad82c1b001efc