Removes the Traitor game mode.
The ```round_autoantag``` and ```latejoin_antags``` are now automatically set based on their original values.
If ```round_autoantag``` is enabled but ```latejoin_antags``` is empty then ```latejoin_antags``` is populated with the core antag ids for the gamemode.
If ```round_autoantag``` is disabled but ```latejoin_antags``` is populated then ```round_autoantag``` is enabled.


Former-commit-id: 03f3e6c3f0ac2e31d60cb7f632c77f6b9a0c40e0