Various tweaks.
A scheduled task that is destroyed now unregisters itself from the source instead of letting it be handled automatically (as this is more efficient).
Makes the Ninja Base a proper area. One should avoid having different names for different areas of the same type.
Floors should now properly unregister the un-wetness task on trigger and destruction.


Former-commit-id: d7374b1faf2e54efb9062fd8443ad0a5f631a6ba