A smoother lobby experience during server start.
Server start initialization now sleeps in a much less aggressive but more efficient manner (similar to how processes sleep). As a result the lobby screen experience is much smoother and allow OOC-chatter, character setup, etc. from the get-go.
The total initialization time does increase some, and as such unit tests don't utilize this at all to complete the process as quickly as possible.

This is basically #13208 but without the attempt to re-organize when and how initialize() is called.


Former-commit-id: d6293e5dff20b683d482d37347b902cfa1a36f7d