Torch Items Part 5: Bits and Bobs
I'll be adding the various items for the Torch in individual
installments so that things aren't stagnant as much. They won't be
anywhere in game 99% of the time, and will all be added later in the
process.

A lot of small stuff that won't be visible by the player other than
having a few extra underwear choices.
Some fixes.
Added new clothing from other corps and organizations for
fluff/event/antag purposes.
A few more things here and there.


Former-commit-id: 14de4b8e4f92cca8c01b6cc4cddce8cea975ec9b