Additional CELLRATE-related tweaks - Mining drills now use CELLRATE correctly. Base power usage is 10kW, with advanced capacitor 5kW, and with super capacitor 3.3kW. - Hardsuits should now use CELLRATE correctly. - Tweaks power usage of hardsuit modules a bit, while i was at it. Many modules have received active power usage, which means they use power when running/on use. The power usages are now in joules/watts, instead of charge units. In the event of future cellrate changes, a similar problem should not occur.
Note: I am currently on a computer that simply isn't powerful enough to run the server to test this stuff. I am opening a PR for review, and to use Travis to compile this stuff, as my compilation time is about 30 minutes at least. I will finish this on Monday, most likely.

Former-commit-id: 0f80a9e89f682c96b08544497e53f586337f4622