Maintenance, junk, and washrooms
The random junk spawner and the maintenance junkifier code both now use the same list of junk.
Restores and adds more potential spots for secret antaggary in maintenance, at least for the industrial minded. Still not much but it's something.
Removes the maintenance random junk spawner as the maintenance initialization code does that as well, as the result is just a metric ton of junk.


Former-commit-id: 5cd2f4f82529081de72c225e4d4cdcf917618e1d