Fix phoron recipes; fix cryo cells claiming bottles are unlabeled
Recipes involving phoron as both catalyst and ingredient were changed by slow-chemistry; they don't fully react any more, because as soon as the catalyst ticks under the limit for the reaction, the reaction stops. Removed phoron catalysts for the two recipes that had phoron as both catalyst and ingredient; recipes which used phoron either only as a catalyst or only as an ingredient are unchanged.

Cryo cells used to require a beaker/bottle be labeled with a pen to show up correctly; bottles labeled with the chemmaster showed up as "No label". Changed this to use the name of the beaker/bottle instead as it's more likely to result in a useful UI. Also changed the loop-over-reagents-and-sum-volume loop, since reagent holders store their total used volume.


Former-commit-id: 78310c97d4d0280104190e78d7b9c8df59405ab6