Moves some overmap stuff into map vars from defines/configs
Like overmap size
And mostly use_overmap flag, I (and other people) keep forgetting to enable it in config and it just screws things up. It would break spacetravel if enabled on a wrong map, and break everything if not enabled on a right map, so config is a pretty bad place for now it now.

Former-commit-id: 22a340d13f33d5ccf27eeb78194db28f222c9ad4