Adds min/max profits reports to roundend summary
Maybe people would care more about money if they get greentext for getting more.

Most of changes is slight cleanup of transactions.
In 99% cases date&time are set to current, so moved that to New().
Amount is now a number instead of acoutantspeak string.
Adds function to apply transaction to account, failed to make it more useful tho.

Also fixes #16127 since I was in the area

Former-commit-id: 01be7b5a56de05a8603e864390e200c5dead1201