Now possible to define the between-Z-level floor per map.
As well as the area for any generated floor turfs.
Utilized by Torch on Z-level 3 as a proof of concept, i.e. most manually placed reinforced floor turfs have been removed, with some minor tweaks to Z-level 2 to make it work.

Tweaks maintenance initialization to not generate vermin and spiderwebs in external areas. Is a little odd.


Former-commit-id: 19934c0c938aefc5044c5f8e8ce9b582d49e53c7