Fix rooms holding pressure when exposed to space.
This happens because edges cease processing when the delta between their
zones is small enough.  This is normally a fine optimization, but when
it results in a room at 4kPa with a window open to space, it breaks
imurshuns.

Two main changes to solve this problem without too much cpu cost:
1) Stop edges from sleeping if one side is a hard vacuum.  This ensures
that a zone doesn't freeze at a low-but-non-zero pressure when touching
hard vacuum.
2) Prevent #1 from causing the edge to stay alive for ages while
pressures asymptotically approach zero as they are repeatedly equalized
but only half is dumped to space. (Would happen if ZoneA---ZoneB---Space
arrangement exists) by detecting when the total amount of air left is
small enough that it would normally sleep anyway, and just setting it to
zero.

The end outcome is that behavior is mostly the same as before, except
when zones have an open path to unsimulated space, they will reach
equilibrium at zero instead of semi-random lowish values.


Former-commit-id: 5c194c8667c1aaa16b2c7a35ca18288391651f3e