Fixes #16806
The latter part of the or check was failing as the proc it was calling was expecting a string and not a list. Checking for the correct species is handled in the earlier part of the or check, and whitelist checking is done when selecting the species for the character in the first place.


Former-commit-id: 8187ff09edadc2b992a8bae2cbd2c63c254b151d