Removes stunbaton beating when set to DISARM.
So, this is made for a few reasons. First off, this prevents officers from beating people unless they REALLY want to. This is probably a good thing and will prevent a lot of accidental beatings.

Secondly, when using an E-gun on stun setting, you need to have it on something other than help. If you then switch to a baton, if it's on disarm, you will begin to beat the shit out of the person. This isn't good game design and sends mixed signals.

Former-commit-id: 3c9ccf21e334326caa91627547f29dff13806f56