Readded the Nanomed and moved the Medbay speaker
Due to comma wanting to keep the nanomed, I moved the (barely used)
medbay intercom. To use it now, you would have to move the freezer a
little, that wont be an issue considering it is barly used. In place I
placed a new nanomed that is now accessable from behind the health
monitor console. Which would make it more accessible for the one
assisting in surgery.


Former-commit-id: 430ef0b212bfc5867b9f4170b2b72f4179eabc98