Changes spess dust event
It used to be a single wave of quasi-meteors that exploded for a bit before dying. There were no further waves until you left and came back. Now waves keep coming ~twice a minute. Dust itself is now a pile of micrometeors - basically space bullets. It wouldn't do a lot of damage to the hull, but can break windows and makes EVA pretty damn dangerous.

Former-commit-id: 5e91784abad037ad4539f33edd9788f9fb89a79f