Nerfs lung damage
At some point it was changed so ANY sort of damage makes lung give you 25% of max health in oxygen loss. This was confusing as  lung wasn't considered 'ruptured' by all the checks.
I bumped it back to ruptured lung, and instead made it give you a small amount of oxyloss just to indicate that something's off.

Former-commit-id: c5c6c2847e79ab9dd8b7cbe12f17471ae0b2742d