Event Tweaks
Activates Solar Storms.

Makes Saferooms and the Bunk Room rad-shielded.

Allows admins to choose whether or not to make an announcement when they
toggle gravity.

Makes it just as likely for some random non-station character to win the
lottery as someone aboard.

Appease Travis

Minor Fixes

Optimization for admin message


Former-commit-id: 4f5a3ab680b5e1900b9b356ed528be66f3a36531