Change EC department accessory to fit on skirts
This removes the knee part of the EC department colour accessory
so that it works on skirts. The department colour thing is
probably part of the proper service uniform, so it should fit.
It's currently impossible for it to fit when it has parts that
attach on parts that don't exist on all of their uniforms.


Former-commit-id: b1c266fb3ae3c887d779bd0115349c9228d06c62