Puts CO's revolver where it belongs
Removes CO's revolver from the closet and puts it in a display case instead.
Tinkers with display cases to allow for anything to be displayed.
Puts lasgun on Bearcat wreck because it was clearly not enough of a Firefly reference already.
Renames CO's revolver to be more appropriate for a diplomatic mission.

Former-commit-id: 52f3ae8bfca8427b210d0a730abe7804a091b940